bower asset packages management added, model image changed to memory, basset deprecated, now using grunt

+++ app/controllers/ObituariesController.php
@@ -1,18 +1,31 @@
 <?php
 
-use Repositories\Interfaces\UserInterface;
-use Repositories\Exceptions\ValidationException as ValidationException;
+namespace Controllers;
+
+use Input;
+use View;
+use Redirect;
+use Authority;
+use Repositories\Interfaces\ObituaryInterface;
+use Repositories\Errors\Exceptions\NotAllowedException as NotAllowedException;
 
 class ObituariesController extends BaseController {
 
 	protected $obituary;
 
 	/**
+	 * The layout that should be used for responses.
+	 */
+	protected $layout = 'layouts.master';
+
+	/**
    * We will use Laravel's dependency injection to auto-magically
    * "inject" our repository instance into our controller
    */
-  public function __construct(UserInterface $obituary){
+  public function __construct(ObituaryInterface $obituary){
     $this->obituary = $obituary;
+
+    $this->beforeFilter('auth', array('except' => 'login'));
   }
 
 	/**
@@ -20,10 +33,13 @@ class ObituariesController extends BaseController {
 	 *
 	 * @return Response
 	 */
-	public function index()
-	{
-		$obituaries = $this->obituary->findAll();
-		return View::make('obituaries.index', compact('obituaries'));
+	public function index(){
+		if( Authority::can('index', 'Obituary') ){
+			$obituaries = $this->obituary->findAll();
+			$this->layout->content = View::make('obituaries.index')->with(compact('obituaries'));;
+			return $this->layout;
+		}
+		throw new NotAllowedException();
 	}
 
 	/**
@@ -31,10 +47,13 @@ class ObituariesController extends BaseController {
 	 *
 	 * @return Response
 	 */
-	public function create()
-	{
-		$obituary = $this->obituary->instance();
-		return View::make('obituaries.create', compact('obituary'));
+	public function create(){
+		if( Authority::can('create', 'Obituary') ){
+			$obituary = $this->obituary->instance();
+			$this->layout->content = View::make('obituaries.create')->with(compact('obituary'));
+			return $this->layout;
+		}
+    throw new NotAllowedException();
 	}
 
 	/**
@@ -42,20 +61,13 @@ class ObituariesController extends BaseController {
 	 *
 	 * @return Response
 	 */
-	public function store()
-	{
-		$input = Input::all();
-		try{
-			$this->obituary->store($input);
-			return Redirect::route('obituaries.index')
-      	->with('success', 'The new obituary has been created');
+	public function store(){
+		if( Authority::can('create', 'Obituary') ){
+			$input = Input::all();
+			$obituary = $this->obituary->store($input);
+			return Redirect::route('obituaries.show', $obituary->id);//->with('success', 'The new obituary has been created');
 		}
-		catch(ValidationException $e){
-			return Redirect::route('obituaries.create')
-		    ->withInput()
-		    ->withErrors($e->getErrors())
-		    ->with('error', 'The obituary was not saved.');
-		}
+    throw new NotAllowedException();
 	}
 
 	/**
@@ -64,15 +76,11 @@ class ObituariesController extends BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function show($id)
-	{
-		try{
-			$obituary = $this->obituary->findById($id);
-			return View::make('obituaries.show', compact('obituary'));
-		}
-		catch(NotFoundException $e){
-			App::abort(404, 'Page not found');
-		}
+	public function show($id){
+		$obituary = $this->obituary->findById($id);
+		//return $this->layout->nest('content', 'obituaries.show', compact('obituary'));
+		$this->layout->content = View::make('obituaries.show')->with(compact('obituary'));
+		return $this->layout;
 	}
 
 	/**
@@ -81,15 +89,13 @@ class ObituariesController extends BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function edit($id)
-	{
-		try{
-			$obituary = $this->obituary->findById($id);
-			return View::make('obituaries.edit', compact('obituary'));
+	public function edit($id){
+		$obituary = $this->obituary->findById($id);
+		if( Authority::can('update', $obituary) ){
+			$this->layout->content = View::make('obituaries.edit')->with(compact('obituary'));
+			return $this->layout;
 		}
-		catch(NotFoundException $e){
-			App::abort(404, 'Page not found');
-		}
+    throw new NotAllowedException();
 	}
 
 	/**
@@ -98,23 +104,13 @@ class ObituariesController extends BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function update($id)
-	{
-		$input = Input::all();
-		try{
-			$this->obituary->update($id, $input);
-			return Redirect::route('obituaries.show', $id)
-				->with('success', 'The new obituary has been created');
+	public function update($id){
+		if( Authority::can('update', 'Obituary') ){
+			$input = Input::all();
+			$obituary = $this->obituary->update($id, $input);
+			return Redirect::route('obituaries.show', $obituary->id);//->with('success', 'The new obituary has been created');
 		}
-		catch(ValidationException $e){
-			return Redirect::route('obituaries.create')
-			->withInput()
-			->withErrors($e->getErrors())
-			->with('error', 'there were validation errors');
-		}
-		catch(NotFoundException $e){
-			App::abort(404, 'Page not found');
-		}
+    throw new NotAllowedException();
 	}
 
 	/**
@@ -123,15 +119,12 @@ class ObituariesController extends BaseController {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function destroy($id)
-	{
-		try{
+	public function destroy($id){
+		if( Authority::can('delete', 'Obituary') ){
 			$this->obituary->destroy($id);
-			return Redirect::route('obituaries.index')->with('success', 'The obituary has been deleted');
+			return Redirect::route('obituaries.index');//->with('success', 'The obituary has been deleted');
 		}
-		catch(NotFoundException $e){
-			App::abort(404, 'Page not found');
-		}
+    throw new NotAllowedException();
 	}
 
 }
